- move autoSelectColorfields into grid file
